feat(listen): wire OAuth single-flight boot gate; split over-cap _agent_exec.py
Half-B of card sac-multi-start-queue-oauth. Concurrent brokered background
spawns share ~/.claude/.credentials.json; their in-container claude refreshes
race the rotating token -> the loser's freshly-rotated (single-use) token is
clobbered -> 401 -> mass re-login. sac never holds token material, so it
COORDINATES access: the listen serializes the credential-sensitive boot window
through a blocking fcntl.flock + bounded refresh-settle.

The natural call-site (_listen/_agent_exec.py) was already 647 lines (over the
512 cap), so split it first (operator-approved):
- _agent_exec_liveness.py: post-ack liveness probe (verbatim extraction).
- _agent_exec_send.py: POST /agents/<name>/send + claude resolver + SSE stream.
- _agent_exec.py: agents_start handler (336 lines) + re-exports (server.py
  import path unchanged); now wraps the background launch in
  run_brokered_launch (single-flight gate). Foreground/one-shot bypass the
  gate (single, interactive, long-lived).

Gate (_credential_refresh_lock.py): unknown expiresAt -> not imminent (no
settle; the flock still serializes the brief window) so creds-less / api-key
setups and the existing listen tests never block. The settle engages only when
expiresAt is known and near expiry, releasing the next waiter once the rotation
lands (expiresAt changes) or SAC_CREDS_SETTLE_SECONDS elapses.

tests/.../_listen/test__credential_refresh_lock.py: AAA, real flock + threads,
yield-fixture env (no mocks/monkeypatch). Verified: imports resolve, all files
<512 lines, 9/9 behavior smoke checks (pytest absent in container venv).

Half-A (bounded-parallel start launcher) and the ~1h in-flight rotation of
already-running agents are tracked as separate follow-ups.
